Add files via upload
In this Python notebook, we will be using a dataset from a competition for classifying feedback as negative or positive feedback, we will encode each word in a feedback sentence into 6d integer vector and for each sentence, we will calculate four unique values using tf-idf techniques stored as 4d vector. Pass these two vectors to a multiple-input neural network.